+++ fastapi_plugins/scheduler.py
@@ -11,11 +11,8 @@ import fastapi
 import pydantic_settings
 import starlette.requests
 
-from .plugin import PluginError
-from .plugin import PluginSettings
-from .plugin import Plugin
-
 from .control import ControlHealthMixin
+from .plugin import Plugin, PluginError, PluginSettings
 from .utils import Annotated
 from .version import VERSION
 
@@ -26,7 +23,7 @@ __all__ = [
 ]
 __author__ = 'madkote <madkote(at)bluewin.ch>'
 __version__ = '.'.join(str(x) for x in VERSION)
-__copyright__ = 'Copyright 2021, madkote'
+__copyright__ = 'Copyright 2025, madkote'
 
 
 class SchedulerError(PluginError):
@@ -142,4 +139,4 @@ async def depends_scheduler(
     return await conn.app.state.AIOJOBS_SCHEDULER()
 
 
-TSchedulerPlugin = Annotated[aiojobs.Scheduler, fastapi.Depends(depends_scheduler)] # noqa E501
+TSchedulerPlugin = Annotated[aiojobs.Scheduler, fastapi.Depends(depends_scheduler)]