+++ tests/conftest.py
@@ -4,6 +4,8 @@
 
 from __future__ import absolute_import
 
+import asyncio
+
 import pytest
 
 
@@ -21,3 +23,10 @@ def pytest_configure(config):
 @pytest.fixture(scope='session')
 def anyio_backend():
     return 'asyncio'
+
+
+@pytest.fixture(scope='session', autouse=True)
+def event_loop(request):
+    loop = asyncio.get_event_loop_policy().new_event_loop()
+    yield loop
+    loop.close()